Create new js file for new functionalities
